tests/requirements.txt
clickzetta-connector
pytest
